requirements.txt
absl-py==2.3.1
agentlace @ git+https://github.com/youliangtan/agentlace.git@cf2c337c5e3694cdbfc14831b239bd657bc4894d
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
array_record==0.7.2
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
chex==0.1.85
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
decorator==4.4.2
distlib==0.4.0
distrax==0.1.5
distro==1.9.0
dlimp @ git+https://github.com/kvablack/dlimp@5edaa4691567873d495633f2708982b42edf1972
dm-tree==0.1.9
easyhid==0.0.10
einops==0.8.1
etils==1.13.0
evdev==1.9.2
exceptiongroup==1.3.0
Farama-Notifications==0.0.4
filelock==3.19.1
flatbuffers==25.2.10
flax==0.7.5
fonttools==4.59.1
fsspec==2025.7.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==0.29.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
identify==2.6.13
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
jax==0.4.20
jaxlib==0.4.20+cuda11.cudnn86
jiter==0.10.0
keras==2.12.0
kiwisolver==1.4.9
libclang==18.1.1
loguru==0.7.3
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.3
moviepy==1.0.3
msgpack==1.1.1
namex==0.1.0
narwhals==2.1.2
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
oauthlib==3.3.1
-e git+ssh://git@github.com/cccedric/octo.git@e454c7a1939163a7e1c70670d34dc504c4825247#egg=octo
omegaconf==2.3.0
openai==1.100.1
opencv-python==4.12.0.88
OpenEXR==3.3.5
opt_einsum==3.4.0
optax==0.1.5
optree==0.17.0
orbax-checkpoint==0.4.4
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
pre-commit==3.3.3
proglog==0.1.12
promise==2.3
protobuf==4.21.12
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynput==1.8.1
pyparsing==3.2.3
pyquaternion==0.9.9
pyrealsense2==2.56.5.9235
python-dateutil==2.9.0.post0
python-xlib==0.33
PyYAML==6.0.2
pyzmq==27.0.1
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.11.4
sentry-sdk==2.35.0
-e git+ssh://git@github.com/cccedric/conrft.git@05b5a4efa2c58e02615713b231e2c0476be8ebe8#egg=serl_launcher&subdirectory=serl_launcher
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-graphics==2021.12.3
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
tensorflow-probability==0.20.1
tensorflow-text==2.15.0
tensorstore==0.1.45
termcolor==3.1.0
tf_keras==2.19.0
tokenizers==0.21.4
toml==0.10.2
toolz==1.0.0
tqdm==4.67.1
transformers==4.55.2
transforms3d==0.4.2
treescope==0.1.10
trimesh==4.7.4
typeguard==2.13.3
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.34.0
wandb==0.21.1
Werkzeug==3.1.3
wrapt==1.14.2
zipp==3.23.0
zmq==0.0.0
